refactor(utils): :building_construction: reorganize the `utils` package structure
