+++ rocrate_validator/requirements/python/__init__.py
@@ -17,12 +17,12 @@ import re
 from pathlib import Path
 from typing import Callable, Optional, Type
 
-import rocrate_validator.log as logging
+from rocrate_validator.utils import log as logging
 from rocrate_validator.models import (LevelCollection, Profile, Requirement,
                                       RequirementCheck, RequirementLevel,
                                       RequirementLoader, Severity,
                                       ValidationContext)
-from rocrate_validator.utils import get_classes_from_file
+from rocrate_validator.utils.python_helpers import get_classes_from_file
 
 # set up logging
 logger = logging.getLogger(__name__)
