+++ rocrate_validator/utils/io_helpers/output/__init__.py
@@ -17,7 +17,7 @@ from typing import Any, Optional, Protocol
 
 from rich.console import Console, ConsoleOptions, RenderResult
 
-import rocrate_validator.log as logging
+from rocrate_validator.utils import log as logging
 
 # set up logging
 logger = logging.getLogger(__name__)
