+++ rocrate_validator/utils/io_helpers/output/json/__init__.py
@@ -1,7 +1,7 @@
 
 from typing import Any, Optional
 
-import rocrate_validator.log as logging
+from rocrate_validator.utils import log as logging
 from rocrate_validator.utils.io_helpers.output import BaseOutputFormatter
 from rocrate_validator.utils.io_helpers.output.json.formatters import (
     ValidationResultJSONOutputFormatter, ValidationResultsJSONOutputFormatter,
