+++ rocrate_validator/utils/rdf.py
@@ -1,0 +1,25 @@
+from rdflib import Graph
+from rocrate_validator.utils import logger
+from rocrate_validator.utils.paths import list_graph_paths
+
+from rocrate_validator import constants
+
+
+def get_full_graph(
+        graphs_dir: str,
+        serialization_format: constants.RDF_SERIALIZATION_FORMATS_TYPES = "turtle",
+        publicID: str = ".") -> Graph:
+    """
+    Get the full graph from the directory
+
+    :param graphs_dir: The directory containing the graphs
+    :param format: The RDF serialization format
+    :param publicID: The public ID
+    :return: The full graph
+    """
+    full_graph = Graph()
+    graphs_paths = list_graph_paths(graphs_dir, serialization_format=serialization_format)
+    for graph_path in graphs_paths:
+        full_graph.parse(graph_path, format="turtle", publicID=publicID)
+        logger.debug("Loaded triples from %s", graph_path)
+    return full_graph
