+++ tests/unit/test_cli_internals.py
@@ -14,7 +14,7 @@
 
 import os
 
-from rocrate_validator import log as logging
+from rocrate_validator.utils import log as logging
 from rocrate_validator.models import (DEFAULT_PROFILES_PATH, LevelCollection,
                                       Profile, ValidationSettings, ValidationStatistics)
